Implement faster poetry env lookup via file system
The CLI was just way too slow. I'm not sure that I've replicated the
poetry activation behaviour exactly, i.e., how it chooses which venv to
activate if more than one are associated with a directory. I've just got
it to activate the last modified environment, which seems like sensible
behaviour.