Fix Windows tests
The tests still don't pass on Windows. Not clear what the solution is.
The paths on Windows obviously have different separators, which means
they hash to a different value, which means the poetry environment will
be created in a different directory.

Maybe we could change the expected directory if we're running on
Windows, or mock the hash function to return a fixed value...